Fix using a custom template not working as expected
This was a lot of work and back and forth with ChatGPT until it finally worked. (I'm no Python developer). But now it works as expected.
You can just copy the template-folder of a publish-output without template option, place it somewhere and then pass a path to this folder to the `--template` argument. Inside this folder you can the modify css and the templates themselves in the views/-folder.